factor data flow
